requirements.txt
aws-cdk-lib==2.131.0
aws-cdk.aws-scheduler-alpha==2.131.0a0
aws-cdk.aws-scheduler-targets-alpha==2.131.0a0
constructs>=10.0.0,<11.0.0
